Move related code to traits
